Keep ViewDragHelper internal state consistent
Code in setDragState() made it possible for the drag state
to become inconsistent with the capturedView state, leading
to a future crash when the state of one (drag state == SETTLING)
led to the assumption about the other (capturedView non-null).

This fix ensures that mCapturedView will only ever be nulled out when
the actual drag state of the helper is IDLE, not the state that someone
attempted to set.

Issue #16797206 [Quantum] App crashes with Null pointer expection, when switching between accounts(seen only once).

CRs-fixed: 776296

Change-Id: I4cc10d7fcee9af1ca14daf9f8e9a4962c255b997